ml/requirements.txt
pandas>=2.1,<2.3
numpy>=1.24,<2.0
scikit-learn>=1.3,<1.6
joblib>=1.3,<2.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
Flask>=3.0,<4.0
